refactor(web): drop legacy /chat/chats route (SOK-795)
Product is early — no compatibility redirect. Delete the chats segment,
move the list shell next to /chat, and remove leftover path assertions.
