--- apps/web/src/app/(app)/chat/chats/loading.tsx
@@ -1,6 +1,0 @@
-import { ChatChatsPageSkeleton } from "@/app/chat/components/chat-chats-loading-view";
-
-/** Sync shell only — no cookies/`connection()` (Instant Nav). */
-export default function ChatChatsLoading() {
-  return <ChatChatsPageSkeleton />;
-}
