+++ apps/web/src/app/(app)/components/__tests__/mobile-app-chrome.test.ts
@@ -94,8 +94,6 @@ describe("mobile-app-chrome", () => {
     it("shows on Welcome home and chats list except rooms and drafts", () => {
       expect(shouldShowMobileBottomNav("/")).toBe(true);
       expect(shouldShowMobileBottomNav("/chat")).toBe(true);
-      // Legacy path redirects; chrome treats it as other-chat (nav via shell).
-      expect(shouldShowMobileBottomNav("/chat/chats")).toBe(true);
       // `welcome=1` used to open the questionnaire, which hid the nav. The
       // param is retired, so a stale link must now behave like bare home.
       expect(
@@ -133,7 +131,6 @@ describe("mobile-app-chrome", () => {
     it("shows brand on Welcome, chats, and every bottom-nav tab root", () => {
       expect(shouldShowMobileBrandLeading("/")).toBe(true);
       expect(shouldShowMobileBrandLeading("/chat")).toBe(true);
-      expect(shouldShowMobileBrandLeading("/chat/chats")).toBe(false);
       // Retired param: falls through to home, so the brand stays.
       expect(
         shouldShowMobileBrandLeading("/", new URLSearchParams("welcome=1")),
